[v0.90][WP-14] Split identity command test responsibilities (Closes #2108)
